refactor(git): promote generic branch helpers from branch-gc into @app/utils/git
