+++ src/utils/git/index.ts
@@ -1,4 +1,4 @@
 export type { GitOptions } from "./core";
-export { createGit, git } from "./core";
+export { BaseNotFoundError, createGit, git } from "./core";
 export * from "./types";
 export * from "./worktree";
